requirements.txt
# ── Stage 1: Data preparation ──────────────────────────────────────────────
# download_data.py
datasets==2.20.0

# preprocess.py
pandas==2.2.2

# ── Stage 2: Fine-tuning ───────────────────────────────────────────────────
# finetune_training.py runs on Kaggle — install separately there.
# Kaggle notebook: https://www.kaggle.com/code/haaziq/cloud-qwen-finetune

# ── Stage 3: Model registry ────────────────────────────────────────────────
# upload_s3.py + register_model.py
boto3==1.34.0
mlflow==2.14.3

# ── Stage 4: Deployment & evaluation ───────────────────────────────────────
# merge_adapter.py + evaluate.py
peft==0.12.0
transformers==4.44.0
accelerate==0.33.0
safetensors==0.4.3

# serve.py (FastAPI inference server)
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0

# evaluate.py
rouge-score==0.1.2
requests==2.32.3

# ── Shared utilities ────────────────────────────────────────────────────────
python-dotenv==1.0.1
rich==13.7.0
tqdm==4.66.4

# ── PyTorch — install separately with the right build for your system ───────
#
#   CPU only (for local inference / WSL2):
#     pip install torch==2.3.0 --index-url https://download.pytorch.org/whl/cpu
#
#   CUDA 12.1 (if you have an NVIDIA GPU):
#     pip install torch==2.3.0 --index-url https://download.pytorch.org/whl/cu121
#
# The Dockerfile installs the CPU build automatically via its own RUN command.
